--- a/hw10_program_optimization/benchmark_test.go
+++ b/hw10_program_optimization/benchmark_test.go
@@ -1,0 +1,15 @@
+package hw10programoptimization
+
+import (
+	"archive/zip"
+	"testing"
+)
+
+func BenchmarkRun(b *testing.B) {
+	r, _ := zip.OpenReader("testdata/users.dat.zip")
+	data, _ := r.File[0].Open()
+
+	b.StartTimer()
+	GetDomainStat(data, "biz")
+	b.StopTimer()
+}