--- a/hw10_program_optimization/benchmark_test.go
+++ b/hw10_program_optimization/benchmark_test.go
@@ -15,57 +15,27 @@ func BenchmarkCaseFromFrozenTest(b *testing.B) {
 	b.ReportAllocs()
 	b.StopTimer()
 
-	for i := 0; i < b.N; i++ {
-		r, _ := zip.OpenReader("testdata/users.dat.zip")
-		data, _ := r.File[0].Open()
+	r, _ := zip.OpenReader("testdata/users.dat.zip")
+	data, _ := r.File[0].Open()
 
-		b.StartTimer()
-		GetDomainStat(data, "biz")
-		b.StopTimer()
+	b.StartTimer()
+	GetDomainStat(data, "biz")
+	b.StopTimer()
 
-		data.Close()
-	}
+	data.Close()
 }
 
 func BenchmarkRandomData(b *testing.B) {
+	b.ReportAllocs()
 	b.StopTimer()
-	lines := generateRandomBuffer(1000)
+	bts := bytes.NewBufferString(generateRandomBuffer(1000))
+	suffix := gofakeit.DomainSuffix()
 	b.StartTimer()
-	b.ReportAllocs()
+
 	for i := 0; i < b.N; i++ {
-		GetDomainStat(bytes.NewBufferString(lines), gofakeit.DomainSuffix())
+		GetDomainStat(bts, suffix)
 	}
 	b.StopTimer()
-}
-
-func BenchmarkExtractEmail(b *testing.B) {
-	b.ReportAllocs()
-	b.StopTimer()
-	for i := 0; i < b.N; i++ {
-		gofakeit.Seed(time.Now().UnixMicro())
-		info := gofakeit.Person()
-
-		if bytes, e := jsoniter.Marshal(*info); e == nil {
-			b.StartTimer()
-			extractEmail(bytes)
-			b.StopTimer()
-		} else {
-			b.FailNow()
-		}
-	}
-}
-
-func BenchmarkHandleEmail(b *testing.B) {
-	b.StopTimer()
-	stat := DomainStat{}
-	for i := 0; i < b.N; i++ {
-		gofakeit.Seed(time.Now().UnixMicro())
-		email := gofakeit.Email()
-
-		b.StartTimer()
-		handleEmail(email, "com", &stat)
-		b.StopTimer()
-	}
 }
 
 func generateRandomBuffer(limit int) string {