--- a/hw10_program_optimization/benchmark_test.go
+++ b/hw10_program_optimization/benchmark_test.go
@@ -2,51 +2,46 @@ package hw10programoptimization
 
 import (
 	"archive/zip"
-	"bytes"
+	"fmt"
 	"strings"
 	"testing"
 	"time"
 
 	"github.com/brianvoe/gofakeit/v6"
-	jsoniter "github.com/json-iterator/go"
 )
 
 func BenchmarkCaseFromFrozenTest(b *testing.B) {
 	b.ReportAllocs()
-	b.StopTimer()
-
-	r, _ := zip.OpenReader("testdata/users.dat.zip")
-	data, _ := r.File[0].Open()
-
-	b.StartTimer()
-	GetDomainStat(data, "biz")
-	b.StopTimer()
-
-	data.Close()
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		b.StopTimer()
+		r, _ := zip.OpenReader("testdata/users.dat.zip")
+		data, _ := r.File[0].Open()
+		b.StartTimer()
+		GetDomainStat(data, "biz")
+		r.Close()
+	}
 }
 
 func BenchmarkRandomData(b *testing.B) {
 	b.ReportAllocs()
-	b.StopTimer()
-	bts := bytes.NewBufferString(generateRandomBuffer(1000))
-	suffix := gofakeit.DomainSuffix()
-	b.StartTimer()
-
+	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
-		GetDomainStat(bts, suffix)
+		b.StopTimer()
+		rdr := strings.NewReader(generateRandomBuffer(1000))
+		suffix := gofakeit.DomainSuffix()
+		b.StartTimer()
+		GetDomainStat(rdr, suffix)
 	}
-	b.StopTimer()
 }
 
 func generateRandomBuffer(limit int) string {
 	gofakeit.Seed(time.Now().UnixMicro())
 	result := strings.Builder{}
 	for i := 0; i < limit; i++ {
-		if bytes, e := jsoniter.Marshal(*gofakeit.Person()); e == nil {
-			result.Write(bytes)
-			result.WriteString("\n")
-		}
+		result.WriteString(fmt.Sprintf(`{"Name":"%s", "Email":"%s"}\n`, gofakeit.Name(), gofakeit.Email()))
 	}
+	line := result.String()
 
-	return result.String()
+	return line
 }