--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -2,11 +2,10 @@ package hw10programoptimization
 
 import (
 	"bufio"
-	"encoding/json"
 	"io"
-	"regexp"
 	"strings"
 )
+import jsoniter "github.com/json-iterator/go"
 
 type User struct {
 	ID       int
@@ -21,53 +20,23 @@ type User struct {
 type DomainStat map[string]int
 
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
-	u, err := getUsers(r)
-	return countDomains(u, err, domain)
-}
-
-func getUsers(r io.Reader) (userCh chan User, e chan error) {
-	userCh = make(chan User)
-	e = make(chan error)
-	go func() {
-		scanner := bufio.NewScanner(r)
-		scanner.Split(bufio.ScanLines)
-		var user User
-		for scanner.Scan() {
-			if err := json.Unmarshal([]byte(scanner.Text()), &user); err != nil {
-				continue
-			}
-			userCh <- user
-		}
-		e <- scanner.Err()
-
-		close(e)
-		close(userCh)
-	}()
-
-	return userCh, e
-}
-
-func countDomains(userCh chan User, e chan error, domain string) (DomainStat, error) {
 	result := make(DomainStat)
-
-	for {
-		select {
-		case u := <-userCh:
-			matched, err := regexp.Match("\\."+domain, []byte(u.Email))
-			if err != nil {
-				return nil, err
-			}
-
-			if matched {
-				num := result[strings.ToLower(strings.SplitN(u.Email, "@", 2)[1])]
-				num++
-				result[strings.ToLower(strings.SplitN(u.Email, "@", 2)[1])] = num
-			}
-		case err := <-e:
-			if err != nil {
-				return nil, err
-			}
-			return result, nil
+	scanner := bufio.NewScanner(r)
+	for scanner.Scan() {
+		email := strings.ToLower(jsoniter.Get(scanner.Bytes(), "Email").ToString())
+		if email == "" || !strings.HasSuffix(email, "."+domain) {
+			continue
 		}
+
+		parts := strings.SplitN(email, "@", 2)
+		if len(parts) != 2 {
+			continue
+		}
+
+		result[parts[1]]++
 	}
+	if scanner.Err() != nil {
+		return nil, scanner.Err()
+	}
+	return result, nil
 }