--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -23,20 +23,28 @@ func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
 	result := make(DomainStat)
 	scanner := bufio.NewScanner(r)
 	for scanner.Scan() {
-		email := strings.ToLower(jsoniter.Get(scanner.Bytes(), "Email").ToString())
-		if email == "" || !strings.HasSuffix(email, "."+domain) {
-			continue
-		}
-
-		parts := strings.SplitN(email, "@", 2)
-		if len(parts) != 2 {
-			continue
-		}
-
-		result[parts[1]]++
+		email := extractEmail(scanner.Bytes())
+		handleEmail(email, domain, &result)
 	}
 	if scanner.Err() != nil {
 		return nil, scanner.Err()
 	}
 	return result, nil
+}
+
+func extractEmail(bytes []byte) string {
+	return strings.ToLower(jsoniter.Get(bytes, "Email").ToString())
+}
+
+func handleEmail(email, domain string, stat *DomainStat) {
+	if !strings.HasSuffix(email, "."+domain) {
+		return
+	}
+
+	parts := strings.SplitN(email, "@", 2)
+	if len(parts) != 2 {
+		return
+	}
+
+	(*stat)[parts[1]]++
 }